# openapi.json
{"openapi":"3.1.0","info":{"title":"Study MCP — Day1","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/whoami":{"get":{"summary":"Whoami","operationId":"whoami_whoami_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/load_material":{"post":{"summary":"Load Material","operationId":"load_material_mcp_tools_load_material_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoadMaterialRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoadMaterialResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/answer_question":{"post":{"summary":"Answer Question","operationId":"answer_question_mcp_tools_answer_question_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerQuestionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerQuestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/make_notes":{"post":{"summary":"Make Notes","operationId":"make_notes_mcp_tools_make_notes_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MakeNotesRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MakeNotesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/generate_quiz":{"post":{"summary":"Generate Quiz","operationId":"generate_quiz_mcp_tools_generate_quiz_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateQuizRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateQuizResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/schedule_quiz":{"post":{"summary":"Schedule Quiz","operationId":"schedule_quiz_mcp_tools_schedule_quiz_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleQuizRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleQuizResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnswerQuestionRequest":{"properties":{"corpus_id":{"type":"string","title":"Corpus Id"},"query":{"type":"string","title":"Query"},"top_k":{"type":"integer","title":"Top K","default":6}},"type":"object","required":["corpus_id","query"],"title":"AnswerQuestionRequest"},"AnswerQuestionResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"citations":{"items":{"additionalProperties":{"type":"string"},"type":"object"},"type":"array","title":"Citations"}},"type":"object","required":["answer","citations"],"title":"AnswerQuestionResponse"},"GenerateQuizRequest":{"properties":{"corpus_id":{"type":"string","title":"Corpus Id"},"duration":{"type":"integer","title":"Duration","default":10},"items":{"type":"integer","title":"Items","default":10},"focus":{"items":{"type":"string"},"type":"array","title":"Focus","default":["definitions","concepts","derivations"]}},"type":"object","required":["corpus_id"],"title":"GenerateQuizRequest"},"GenerateQuizResponse":{"properties":{"quiz_id":{"type":"string","title":"Quiz Id"},"items":{"items":{"$ref":"#/components/schemas/QuizItem"},"type":"array","title":"Items"}},"type":"object","required":["quiz_id","items"],"title":"GenerateQuizResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoadMaterialRequest":{"properties":{"file_id":{"type":"string","title":"File Id","description":"Arbitrary ID"},"type":{"type":"string","title":"Type","description":"One of: \"pdf\",\"pptx\",\"docx\""},"range":{"type":"string","title":"Range","description":"e.g., \"slides 1-5\" or \"pages 3-10\""},"local_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Local Path"}},"type":"object","required":["file_id","type","range"],"title":"LoadMaterialRequest"},"LoadMaterialResponse":{"properties":{"corpus_id":{"type":"string","title":"Corpus Id"},"chunks_indexed":{"type":"integer","title":"Chunks Indexed"},"sources":{"items":{"type":"string"},"type":"array","title":"Sources"}},"type":"object","required":["corpus_id","chunks_indexed","sources"],"title":"LoadMaterialResponse"},"MakeNotesRequest":{"properties":{"corpus_id":{"type":"string","title":"Corpus Id"},"style":{"type":"string","title":"Style","default":"concise"},"export_pdf":{"type":"boolean","title":"Export Pdf","default":true}},"type":"object","required":["corpus_id"],"title":"MakeNotesRequest"},"MakeNotesResponse":{"properties":{"notes_md":{"type":"string","title":"Notes Md"},"pdf_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Url"}},"type":"object","required":["notes_md"],"title":"MakeNotesResponse"},"QuizItem":{"properties":{"qtype":{"type":"string","title":"Qtype"},"question":{"type":"string","title":"Question"},"options":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Options"},"answer":{"type":"string","title":"Answer"},"source_id":{"type":"string","title":"Source Id"},"rationale":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rationale"}},"type":"object","required":["qtype","question","answer","source_id"],"title":"QuizItem"},"SchedulePlan":{"properties":{"mode":{"type":"string","title":"Mode"},"end_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Date"},"days":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Days"},"window":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Window"}},"type":"object","required":["mode"],"title":"SchedulePlan"},"ScheduleQuizRequest":{"properties":{"quiz_id":{"type":"string","title":"Quiz Id"},"plan":{"$ref":"#/components/schemas/SchedulePlan"},"title":{"type":"string","title":"Title","default":"Pop Quiz"},"tz":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tz"},"confirm":{"type":"boolean","title":"Confirm","default":true}},"type":"object","required":["quiz_id","plan"],"title":"ScheduleQuizRequest"},"ScheduleQuizResponse":{"properties":{"events":{"items":{"$ref":"#/components/schemas/ScheduledEvent"},"type":"array","title":"Events"},"ics_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ics Url"},"preview_only":{"type":"boolean","title":"Preview Only","default":false}},"type":"object","required":["events"],"title":"ScheduleQuizResponse"},"ScheduledEvent":{"properties":{"start":{"type":"string","title":"Start"},"end":{"type":"string","title":"End"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"}},"type":"object","required":["start","end","title","description"],"title":"ScheduledEvent"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}